Fix streamer adapter thread restart
There was noticed an issue with the streamer thread
restart. This issue is happening when streaming activity
was stopped and the started again in scope of the same
session. In case when `StopActivity()` is called then
`StreamerAdapter::Streamer::exitThreadMain` is called on
stop which causes `messages_` queue to go to
shutdown to avoid new messages be added into the queue.
This also marks the queue as shutting down.
After that, when the same session adapter `StartActivity()`
called and thread restarted, messages queue might be still
shutting down.
To avoid that, `Reset()` will be called on thread start in
order to reset shutting down flag. Also, exitThreadMain()
method was replaced with `Stop()` method to properly wait
for streaming thread to be completely stopped.